% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_analysis_functions.R
\name{blacksheep}
\alias{blacksheep}
\title{Run the entire Blacksheep Function from Start to finish}
\usage{
blacksheep(counttable, metatable, analyze_negative_outliers = FALSE,
  aggregate_features = FALSE, feature_delineator = "\\\\.",
  fdrcutoffvalue = 0.1, outfilepath = getwd())
}
\arguments{
\item{counttable}{the count table to analyze}

\item{metatable}{the metatable with each column representing a comparison to
perform.}

\item{analyze_negative_outliers}{DEFAULT: FALSE; Toggle the analysis of
outliers in the negative direction as well. Will lead to the output of
the outlier table containing "-1" values, in addition to negative outputs
for boundaries and aggregate tables (if applicable)}

\item{aggregate_features}{DEFAULT: FALSE; Toggle the Aggregate feature, which
will aggregate features in your table based on the given delineator.
Aggregation will output an aggregate table that counts the number of
outliers per feature, and also a fraction table that show the number of
outliers / number of candidates (which excludes missing values)}

\item{feature_delineator}{DEFAULT: <"\\."> What character delineates the
separation between primary and secondary features. NOTE: to use proper
R syntax with escape characters if necessary
Ex) Protein1.Phosphosite1 uses "\\." to aggregate on Protein1}

\item{fdrcutoffvalue}{DEFAULT: 0.1; The FDR value for significance}

\item{outfilepath}{the full string path to where the file should output to,
DEFAULT is current working directory}
}
\value{
outputs a pdf with the heatmap in the current working directory
}
\description{
Run the entire Blacksheep Function from Start to finish
}
\examples{

data("sample_values")
data("sample_annotations")

blacksheep(counttable = sample_values, metatable = sample_annotations,
    analyze_negative_outliers = TRUE, aggregate_features = TRUE,
    feature_delineator = "\\\\.", fdrcutoffvalue = 0.1,
    outfilepath = getwd())
}
\keyword{outliers}
