% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_analysis_functions.R
\name{comparison_groupings}
\alias{comparison_groupings}
\title{Create all of the groups based on the input metadata}
\usage{
comparison_groupings(comptable)
}
\arguments{
\item{comptable}{table where each column will have comparisons drawn from it}
}
\value{
a list with each of the groups as an entry in the list
    NOTE - this list will be ncol*2 long where ncol is the number comparisons
}
\description{
Create all of the groups based on the input metadata
}
\examples{
data("sample_annotations")
groupings = comparison_groupings(sample_annotations)
}
\keyword{outliers}
