% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_plotting_functions.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Plot out a heatmap}
\usage{
create_heatmap(counttab = counttab,
    colmetatable = NULL, colannotationlist = NULL,
    colclusterparam = FALSE, rowclusterparam = FALSE,
    nameparam, write_out_plot = FALSE, pdfoutfile = pdfoutfile)
}
\arguments{
\item{counttab}{table with counts, samples -x-axis, features -y-axis}

\item{colmetatable}{the metatable containing information for the columns}

\item{colannotationlist}{annotation table for columns, based off colmetatable}

\item{colclusterparam}{cluster the columns?}

\item{rowclusterparam}{cluster the rows?}

\item{nameparam}{the title on the heatmap}

\item{write_out_plot}{DEFAULT: FALSE write out the plot to <pdfoutfile>}

\item{pdfoutfile}{the full path to the pdf to write out to}
}
\value{
prints a pdf heatmap out to the designated outpath
}
\description{
Plot out a heatmap
}
\examples{
data("sample_values")
counttab = sample_values
pdfoutfile = paste(getwd(), "/heatmap_out.pdf", sep = "")
nameparam = "testplot"

create_heatmap(counttab = counttab,
    colmetatable = NULL,
    colannotationlist = NULL,colclusterparam = FALSE,
    rowclusterparam = FALSE, nameparam, write_out_plot = FALSE,
    pdfoutfile= pdfoutfile)
}
\keyword{outliers}
