% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_analysis_functions.R
\name{outlier_heatmap}
\alias{outlier_heatmap}
\title{With the grouptablist generated by count_outliers - run through and run a
    fisher exact test to get the p.value for the difference in outlier count
    for each feature in each of your comparisons}
\usage{
outlier_heatmap(outlier_analysis_out, analysis_num = NULL, counttab,
  metatable, fdrcutoffvalue = 0.1, write_out_plot = FALSE,
  outfilepath = getwd())
}
\arguments{
\item{outlier_analysis_out}{the full outlier_analysis data objet}

\item{analysis_num}{DEFAULT: NULL; if you only want to plot the heatmap for
a particular analysis, enter number of that analysis}

\item{counttab}{the raw data before outlier analysis}

\item{metatable}{the complete metatable that was used to generate the
comparisons, will be used for annotation of the heatmap}

\item{fdrcutoffvalue}{DEFAULT: 0.1; The FDR value for significance}

\item{write_out_plot}{DEFAULT: FALSE; write out the plot to <outfilepath>}

\item{outfilepath}{the full string path to where the file should output to,
DEFAULT is current working directory}
}
\value{
outputs a pdf with the heatmap in the current working directory
}
\description{
With the grouptablist generated by count_outliers - run through and run a
    fisher exact test to get the p.value for the difference in outlier count
    for each feature in each of your comparisons
}
\examples{

data("sample_values")
reftable_function_out = make_outlier_table(sample_values)
outliertab = reftable_function_out$outliertab

data("sample_annotations")
groupings = comparison_groupings(sample_annotations)

grouptablist = count_outliers(groupings, outliertab)

outlier_analysis_out = outlier_analysis(grouptablist)

metatable = sample_annotations
counttab = sample_values

outlier_heatmap(outlier_analysis_out, analysis_num = NULL,
    counttab, metatable, fdrcutoffvalue = 0.1)
}
\keyword{outliers}
