% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_annotationdata}
\alias{sample_annotationdata}
\title{sample_annotationdata}
\format{A data frame with 76 rows and 6 variables:
\describe{
    \item{PAM50_Her2}{The binary PAM50 Her2 classification for each sample}
    \item{PAM50_Basal}{The binary PAM50 Basal classification for each sample}
    \item{PAM50_LumA}{The binary PAM50 LumA classification for each sample}
    \item{PAM50_LumB}{The binary PAM50 LumB classification for each sample}
    \item{ER_Status}{The ER Status classification for each sample}
    \item{PR_Status}{The PR Status classification for each sample}
    ...
}}
\source{
\url{https://cptac-data-portal.georgetown.edu/cptac/s/S029}
}
\usage{
sample_annotationdata
}
\description{
Example annotation data for Outlier analysis.
This example data is a subset of the data used in the CPTAC3 Breast Cancer
exploration study: (doi: 10.1038/nature18003).
Each row corresponds to a sample and each column is an binary annotation for
that sample.
}
\keyword{datasets}
