% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_phosphodata}
\alias{sample_phosphodata}
\title{sample_phosphodata}
\format{A data frame with 15532 rows and 76 variables:
\describe{
    \item{TCGA-A2-A0CM}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0D2}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0EQ}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0EV}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0EX}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0EY}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0SW}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0SX}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0T3}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0T6}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0YC}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0YD}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0YF}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0YG}{phosphoprotein levels for each gene}
    \item{TCGA-A2-A0YM}{phosphoprotein levels for each gene}
    \item{TCGA-A7-A0CE}{phosphoprotein levels for each gene}
    \item{TCGA-A7-A0CJ}{phosphoprotein levels for each gene}
    \item{TCGA-A7-A13F}{phosphoprotein levels for each gene}
    \item{TCGA-A8-A06N}{phosphoprotein levels for each gene}
    \item{TCGA-A8-A06Z}{phosphoprotein levels for each gene}
    \item{TCGA-A8-A076}{phosphoprotein levels for each gene}
    \item{TCGA-A8-A079}{phosphoprotein levels for each gene}
    \item{TCGA-A8-A08Z}{phosphoprotein levels for each gene}
    \item{TCGA-A8-A09G}{phosphoprotein levels for each gene}
    \item{TCGA-AN-A04A}{phosphoprotein levels for each gene}
    \item{TCGA-AN-A0AJ}{phosphoprotein levels for each gene}
    \item{TCGA-AN-A0AL}{phosphoprotein levels for each gene}
    \item{TCGA-AN-A0AM}{phosphoprotein levels for each gene}
    \item{TCGA-AN-A0FK}{phosphoprotein levels for each gene}
    \item{TCGA-AN-A0FL}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A03O}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0J6}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0J9}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0JC}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0JE}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0JJ}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0JL}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A0JM}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A126}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A12B}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A12D}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A12E}{phosphoprotein levels for each gene}
    \item{TCGA-AO-A12F}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A0TR}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A0TT}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A0TV}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A0TX}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A0U4}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A1AP}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A1AS}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A1AV}{phosphoprotein levels for each gene}
    \item{TCGA-AR-A1AW}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0AV}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0BV}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0C1}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0C7}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0DD}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0DG}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0E1}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A0E9}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A18N}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A18Q}{phosphoprotein levels for each gene}
    \item{TCGA-BH-A18U}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A12L}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A12T}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A12U}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A12V}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A12Z}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A130}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A131}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A134}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A135}{phosphoprotein levels for each gene}
    \item{TCGA-C8-A138}{phosphoprotein levels for each gene}
    \item{TCGA-D8-A142}{phosphoprotein levels for each gene}
    \item{TCGA-E2-A154}{phosphoprotein levels for each gene}
    \item{TCGA-E2-A158}{phosphoprotein levels for each gene}
}}
\source{
\url{https://cptac-data-portal.georgetown.edu/cptac/s/S029}
}
\usage{
sample_phosphodata
}
\description{
Example phosphoprotein data for Outlier analysis
This example data is a subset of the data used in the CPTAC3 Breast Cancer
exploration study: (doi: 10.1038/nature18003).
Each row corresponds to a phosphoprotein site, and each column is a sample.
The values within the table are normalized massspec phosphoprotein values.
}
\keyword{datasets}
