% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_rnadata}
\alias{sample_rnadata}
\title{sample_rnadata}
\format{A data frame with 4317 rows and 76 variables:
\describe{
    \item{TCGA-A2-A0CM}{RNA levels for each gene}
    \item{TCGA-A2-A0D2}{RNA levels for each gene}
    \item{TCGA-A2-A0EQ}{RNA levels for each gene}
    \item{TCGA-A2-A0EV}{RNA levels for each gene}
    \item{TCGA-A2-A0EX}{RNA levels for each gene}
    \item{TCGA-A2-A0EY}{RNA levels for each gene}
    \item{TCGA-A2-A0SW}{RNA levels for each gene}
    \item{TCGA-A2-A0SX}{RNA levels for each gene}
    \item{TCGA-A2-A0T3}{RNA levels for each gene}
    \item{TCGA-A2-A0T6}{RNA levels for each gene}
    \item{TCGA-A2-A0YC}{RNA levels for each gene}
    \item{TCGA-A2-A0YD}{RNA levels for each gene}
    \item{TCGA-A2-A0YF}{RNA levels for each gene}
    \item{TCGA-A2-A0YG}{RNA levels for each gene}
    \item{TCGA-A2-A0YM}{RNA levels for each gene}
    \item{TCGA-A7-A0CE}{RNA levels for each gene}
    \item{TCGA-A7-A0CJ}{RNA levels for each gene}
    \item{TCGA-A7-A13F}{RNA levels for each gene}
    \item{TCGA-A8-A06N}{RNA levels for each gene}
    \item{TCGA-A8-A06Z}{RNA levels for each gene}
    \item{TCGA-A8-A076}{RNA levels for each gene}
    \item{TCGA-A8-A079}{RNA levels for each gene}
    \item{TCGA-A8-A08Z}{RNA levels for each gene}
    \item{TCGA-A8-A09G}{RNA levels for each gene}
    \item{TCGA-AN-A04A}{RNA levels for each gene}
    \item{TCGA-AN-A0AJ}{RNA levels for each gene}
    \item{TCGA-AN-A0AL}{RNA levels for each gene}
    \item{TCGA-AN-A0AM}{RNA levels for each gene}
    \item{TCGA-AN-A0FK}{RNA levels for each gene}
    \item{TCGA-AN-A0FL}{RNA levels for each gene}
    \item{TCGA-AO-A03O}{RNA levels for each gene}
    \item{TCGA-AO-A0J6}{RNA levels for each gene}
    \item{TCGA-AO-A0J9}{RNA levels for each gene}
    \item{TCGA-AO-A0JC}{RNA levels for each gene}
    \item{TCGA-AO-A0JE}{RNA levels for each gene}
    \item{TCGA-AO-A0JJ}{RNA levels for each gene}
    \item{TCGA-AO-A0JL}{RNA levels for each gene}
    \item{TCGA-AO-A0JM}{RNA levels for each gene}
    \item{TCGA-AO-A126}{RNA levels for each gene}
    \item{TCGA-AO-A12B}{RNA levels for each gene}
    \item{TCGA-AO-A12D}{RNA levels for each gene}
    \item{TCGA-AO-A12E}{RNA levels for each gene}
    \item{TCGA-AO-A12F}{RNA levels for each gene}
    \item{TCGA-AR-A0TR}{RNA levels for each gene}
    \item{TCGA-AR-A0TT}{RNA levels for each gene}
    \item{TCGA-AR-A0TV}{RNA levels for each gene}
    \item{TCGA-AR-A0TX}{RNA levels for each gene}
    \item{TCGA-AR-A0U4}{RNA levels for each gene}
    \item{TCGA-AR-A1AP}{RNA levels for each gene}
    \item{TCGA-AR-A1AS}{RNA levels for each gene}
    \item{TCGA-AR-A1AV}{RNA levels for each gene}
    \item{TCGA-AR-A1AW}{RNA levels for each gene}
    \item{TCGA-BH-A0AV}{RNA levels for each gene}
    \item{TCGA-BH-A0BV}{RNA levels for each gene}
    \item{TCGA-BH-A0C1}{RNA levels for each gene}
    \item{TCGA-BH-A0C7}{RNA levels for each gene}
    \item{TCGA-BH-A0DD}{RNA levels for each gene}
    \item{TCGA-BH-A0DG}{RNA levels for each gene}
    \item{TCGA-BH-A0E1}{RNA levels for each gene}
    \item{TCGA-BH-A0E9}{RNA levels for each gene}
    \item{TCGA-BH-A18N}{RNA levels for each gene}
    \item{TCGA-BH-A18Q}{RNA levels for each gene}
    \item{TCGA-BH-A18U}{RNA levels for each gene}
    \item{TCGA-C8-A12L}{RNA levels for each gene}
    \item{TCGA-C8-A12T}{RNA levels for each gene}
    \item{TCGA-C8-A12U}{RNA levels for each gene}
    \item{TCGA-C8-A12V}{RNA levels for each gene}
    \item{TCGA-C8-A12Z}{RNA levels for each gene}
    \item{TCGA-C8-A130}{RNA levels for each gene}
    \item{TCGA-C8-A131}{RNA levels for each gene}
    \item{TCGA-C8-A134}{RNA levels for each gene}
    \item{TCGA-C8-A135}{RNA levels for each gene}
    \item{TCGA-C8-A138}{RNA levels for each gene}
    \item{TCGA-D8-A142}{RNA levels for each gene}
    \item{TCGA-E2-A154}{RNA levels for each gene}
    \item{TCGA-E2-A158}{RNA levels for each gene}
}}
\source{
\url{https://cptac-data-portal.georgetown.edu/cptac/s/S029}
}
\usage{
sample_rnadata
}
\description{
Example RNA data for Outlier analysis
This example data is a subset of the data used in the CPTAC3 Breast Cancer
exploration study: (doi: 10.1038/nature18003).
Each row corresponds to a gene, and each column is a sample. The values
within the table are normalized transcript counts.
}
\keyword{datasets}
