% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_values}
\alias{sample_values}
\title{Sample_values Data}
\format{A data frame with 18 rows and 21 variables:
\describe{
    \item{geneA.site1}{Gene A, example site 1}
    \item{geneB.site1}{Gene B, example site 1}
    ...
}}
\usage{
sample_values
}
\description{
Example count data for Outlier analysis
}
\keyword{datasets}
