% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_analysis_functions.R
\name{count_outliers}
\alias{count_outliers}
\title{Count up the outlier information for each of the groups you have made}
\usage{
count_outliers(groupings, outliertab, analyze_negative_outliers = FALSE)
}
\arguments{
\item{groupings}{table generated by the comparison_groupings function}

\item{outliertab}{outlier table generated by make_outlier_table}

\item{analyze_negative_outliers}{DEFAULT: FALSE; analyze negative outliers in
addition to positive outliers, NOTE - this MUST BE SET TO TRUE if using
an aggregate negative table to get a result}
}
\value{
the tabulated information of outliers per group
}
\description{
Count up the outlier information for each of the groups you have made
}
\examples{

data("sample_values")
reftable_function_out = make_outlier_table(sample_values)
outliertab = reftable_function_out$outliertab

data("sample_annotations")
groupings = comparison_groupings(sample_annotations)

grouptablist = count_outliers(groupings, outliertab,
    analyze_negative_outliers = FALSE)
}
\keyword{outliers}
