% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_analysis_functions.R
\name{outlier_analysis}
\alias{outlier_analysis}
\title{With the grouptablist generated by count_outliers - run through and run a
    fisher exact test to get the p.value for the difference in outlier count
    for each feature in each of your comparisons}
\usage{
outlier_analysis(grouptablist, write_out_tables = FALSE,
  outfilepath = getwd())
}
\arguments{
\item{grouptablist}{table generated by the count_outliers function. NOTE that
the inputted grouptablist will be deciphered to determine its content.
This means that user decides to input the outliertab or aggregate tab,
and the output will analyze according to what positive and negative
information is contained within the table}

\item{write_out_tables}{DEFAULT: FALSE; utility in function to write out
each of the analyses to a separate table to whereever <outfilepath> is
specfied.}

\item{outfilepath}{the full string path to where the file should output to,
DEFAULT is current working directory}
}
\value{
the analysis table with p.value, fdr, and raw data per comparison
}
\description{
With the grouptablist generated by count_outliers - run through and run a
    fisher exact test to get the p.value for the difference in outlier count
    for each feature in each of your comparisons
}
\examples{

data("sample_values")
head(sample_values)
reftable_function_out = make_outlier_table(sample_values)
outliertab = reftable_function_out$outliertab

data("sample_annotations")
groupings = comparison_groupings(sample_annotations)

grouptablist = count_outliers(groupings, outliertab)

outlier_analysis_out = outlier_analysis(grouptablist)
}
\keyword{outliers}
